update session footer, setup add exercise at end of workout
