+++ client/src/features/sessions/SessionExercise.tsx
@@ -6,6 +6,8 @@ import Link from 'next/link';
 import { GetSessionExerciseResponse } from './types';
 import { MyTimer } from 'features/timers/Timer';
 import { TimerType } from 'features/timers/types';
+import { BsArrowLeftShort, BsArrowRightShort } from 'react-icons/bs';
+import { SessionFooter } from './SessionFooter';
 
 interface Props {
   sessionId: string;
@@ -17,16 +19,6 @@ export const SessionExercise = ({ sessionId, exerciseId, data }: Props) => {
   return (
     <main className="pb-20 pt-16">
       <div className="px-3 py-5 bg-gray-50" />
-
-      {!!data.exercise.video && (
-        <div className="pb-5 w-full">
-          <video
-            controls
-            src={`https://cdn.trckd.ca/${data.exercise.video}`}
-            className="w-full"
-          />
-        </div>
-      )}
 
       <div className="flex justify-evenly pt-5">
         <p>Reps</p>
@@ -46,59 +38,11 @@ export const SessionExercise = ({ sessionId, exerciseId, data }: Props) => {
         workoutId={data.exercise.workout_id}
       />
 
-      <Footer nextEx={data.next} prevEx={data.prev} sessionId={sessionId} />
-    </main>
-  );
-};
-
-interface SessionFooterProps {
-  sessionId: string;
-  nextEx: {
-    order: {
-      exercise_id: string;
-    };
-  };
-  prevEx: {
-    order: {
-      exercise_id: string;
-    };
-  };
-}
-
-const Footer = ({ prevEx, nextEx, sessionId }: SessionFooterProps) => {
-  const [timerType, setTimerType] = useState<TimerType>(null);
-  const time = new Date();
-  return (
-    <div>
-      <MyTimer
-        expiryTimestamp={time}
-        timerType={timerType}
-        setTimerType={setTimerType}
+      <SessionFooter
+        nextEx={data.next}
+        prevEx={data.prev}
+        sessionId={sessionId}
       />
-
-      <div className="flex justify-between py-3 px-3 fixed bottom-0 bg-white left-0 right-0 ">
-        {!!prevEx.order.exercise_id && (
-          <Link
-            href={`/sessions/${sessionId}/exercises/${prevEx.order.exercise_id}`}
-          >
-            <Button className="w-full">Previous</Button>
-          </Link>
-        )}
-
-        {!!nextEx.order.exercise_id && (
-          <Link
-            href={`/sessions/${sessionId}/exercises/${nextEx.order.exercise_id}`}
-          >
-            <Button className="w-full">Next</Button>
-          </Link>
-        )}
-
-        {!nextEx.order.exercise_id && (
-          <Link href={`/sessions/${sessionId}/end`}>
-            <Button className="w-full">Finish Workout</Button>
-          </Link>
-        )}
-      </div>
-    </div>
+    </main>
   );
 };
